requirements.txt
click>=8.0.0
rich>=13.0.0
sqlalchemy>=2.0.0
pydantic>=2.0.0
plyer>=2.1.0
pyyaml>=6.0
python-dateutil>=2.8.0
